connector_search_engine,connector_elasticsearch: prepare test refactoring
For now only move elastic test into a common class
The idea is to have the same test for all adapter
as all adapter should have the same behaviour
